# contracts/openapi/core-api-v1.yaml
openapi: 3.1.0
info:
  title: M4Trust Core API
  version: 1.0.0
  description: Reviewed design contract for the same-origin Spring API consumed by the M4Trust frontend.
servers:
  - url: /api/v1
    description: Same-origin public API v1
...
paths:
...
  /deals:
    post:
      operationId: createDeal
      tags: [Deals]
      summary: Create a Deal in DRAFT status
      description: >
        Creates a Deal for the active legal entity, which becomes the initial
        participant. The title is trimmed and must remain non-blank. The optional
        description is stored as null when omitted. No idempotency key is used in
        this slice.
      security:
        - SessionCookie: []
          CsrfToken: []
      parameters:
        - $ref: '#/components/parameters/LegalEntityContext'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealRequest'
      responses:
        '201':
          description: Deal created in DRAFT status.
          headers:
            Location:
              description: Same-origin path of the created Deal.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealDetail'
        '400':
          $ref: '#/components/responses/MalformedRequest'
        '401':
          $ref: '#/components/responses/SessionRequired'
        '403':
          $ref: '#/components/responses/DealScopedMutationForbidden'
        '404':
          $ref: '#/components/responses/LegalEntityNotFoundOrHidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
    get:
      operationId: listDeals
      tags: [Deals]
      summary: List Deals visible to the active legal entity
      description: >
        Returns only Deals in which the active legal entity is a participant. Results
        use stable page metadata. The default sort is createdAt,desc; the backend may
        add a stable identifier tie-breaker without exposing persistence field names.
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/LegalEntityContext'
        - $ref: '#/components/parameters/DealStatusFilter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/DealSort'
      responses:
        '200':
          description: Page of Deals visible to the active legal entity; items is never null.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealPage'
        '400':
          $ref: '#/components/responses/MalformedRequest'
        '401':
          $ref: '#/components/responses/SessionRequired'
        '403':
          $ref: '#/components/responses/LegalEntityAccessDenied'
        '404':
          $ref: '#/components/responses/LegalEntityNotFoundOrHidden'
        '422':
          $ref: '#/components/responses/ValidationFailed'
...
  /deals/{dealId}/extraction-review:
    get:
      operationId: getDealExtractionReview
      tags: [Deal Rule Sets]
      summary: Get the participant-readable current extraction review projection
      description: >
        Returns the canonical extracted rules and their review input for the current
        REVIEW_REQUIRED analysis to every Deal participant. This is review material,
        not ratification or contractual consent. A visible Deal without a current
        reviewable analysis returns 409 DEAL_STATE_CONFLICT; a hidden Deal remains
        non-disclosing.
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/DealId'
        - $ref: '#/components/parameters/LegalEntityContext'
      responses:
        '200':
          description: Current review projection visible to the active Deal participant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealExtractionReview'
        '400':
          $ref: '#/components/responses/MalformedRequest'
        '401':
          $ref: '#/components/responses/SessionRequired'
        '403':
          $ref: '#/components/responses/LegalEntityAccessDenied'
        '404':
          $ref: '#/components/responses/DealOrLegalEntityNotFoundOrHidden'
        '409':
          $ref: '#/components/responses/DealReviewConflict'
  /deals/{dealId}/extraction-review/accept:
    post:
      operationId: acceptDealExtractionReview
      tags: [Deal Rule Sets]
      summary: Accept a reviewed extraction as an immutable rule-set version
      description: >
        Initiator-only bulk review action. The required target analysisId must still
        be the current REVIEW_REQUIRED analysis for the current document under the
        Deal lock. Idempotency-Key is required: replaying the same canonical request
        returns the originally created RuleSetVersion, while reuse for a different
        request returns 409 IDEMPOTENCY_KEY_REUSED. A successful action atomically
        creates one immutable RuleSetVersion, marks the analysis ACCEPTED, and moves
        the Deal current rule-set pointer. It is not ratification or commercial consent.
      security:
        - SessionCookie: []
          CsrfToken: []
      parameters:
        - $ref: '#/components/parameters/DealId'
        - $ref: '#/components/parameters/LegalEntityContext'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptExtractionReviewRequest'
      responses:
        '201':
          description: Immutable RuleSetVersion created; idempotent replay returns this original version.
          headers:
            Location:
              description: Same-origin path of the created immutable rule-set version.
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleSetVersion'
        '400':
          $ref: '#/components/responses/MalformedRequest'
        '401':
          $ref: '#/components/responses/SessionRequired'
        '403':
          $ref: '#/components/responses/DealReviewAcceptanceForbidden'
        '404':
          $ref: '#/components/responses/DealOrLegalEntityNotFoundOrHidden'
        '409':
          $ref: '#/components/responses/DealReviewAcceptanceConflict'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /deals/{dealId}/rule-set-versions:
    get:
      operationId: listDealRuleSetVersions
      tags: [Deal Rule Sets]
      summary: List immutable rule-set version history
      description: >
        Returns all immutable accepted rule-set versions for the Deal to every
        participant, including superseded historical versions. No version may be
        edited, deleted, or inferred as current from position in this list.
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/DealId'
        - $ref: '#/components/parameters/LegalEntityContext'
      responses:
        '200':
          description: Immutable rule-set history visible to the active Deal participant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleSetVersionHistory'
        '400':
          $ref: '#/components/responses/MalformedRequest'
        '401':
          $ref: '#/components/responses/SessionRequired'
        '403':
          $ref: '#/components/responses/LegalEntityAccessDenied'
        '404':
          $ref: '#/components/responses/DealOrLegalEntityNotFoundOrHidden'
  /deals/{dealId}/rule-set-versions/{ruleSetVersionId}:
    get:
      operationId: getDealRuleSetVersion
      tags: [Deal Rule Sets]
      summary: Get one immutable rule-set version
      security:
        - SessionCookie: []
      parameters:
        - $ref: '#/components/parameters/DealId'
        - $ref: '#/components/parameters/RuleSetVersionId'
        - $ref: '#/components/parameters/LegalEntityContext'
      responses:
        '200':
          description: Immutable rule-set version visible to the active Deal participant.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuleSetVersion'
        '400':
          $ref: '#/components/responses/MalformedRequest'
        '401':
          $ref: '#/components/responses/SessionRequired'
        '403':
          $ref: '#/components/responses/LegalEntityAccessDenied'
        '404':
          $ref: '#/components/responses/RuleSetVersionNotFoundOrHidden'
...
components:
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-M4TRUST_SESSION
      description: >
        Opaque server-side session cookie. Production uses this HttpOnly, Secure,
        SameSite=Lax, Path=/ cookie without a Domain attribute. Local development
        uses the documented non-Secure M4TRUST_SESSION variant. Client JavaScript
        must never read, copy, or log the cookie.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: >
        CSRF token returned by GET /security/csrf and required on unsafe requests.
        It is not an authentication credential or a session identifier.
...
  parameters:
...
    LegalEntityContext:
      name: X-M4Trust-Legal-Entity-Id
      in: header
      required: true
      description: >
        Client-selected legal entity context. This header is neither an authentication
        credential nor proof of authorization. It is verified server-side against
        membership and, for Deal operations, participation. On a legal entity detail
        path it must equal the legalEntityId path parameter. Missing, malformed, or
        mismatched values produce LEGAL_ENTITY_ACCESS_DENIED.
      schema:
        type: string
        format: uuid
    DealId:
      name: dealId
      in: path
      required: true
      description: Public UUID of the requested Deal.
      schema:
        type: string
        format: uuid
...
    RuleSetVersionId:
      name: ruleSetVersionId
      in: path
      required: true
      description: Immutable RuleSetVersion identifier.
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >
        UUID key supplied by the client for one idempotent operation. Keep the same
        key when retrying the same canonical request; never use it for a different
        request. A different request with an already-used key returns 409
        IDEMPOTENCY_KEY_REUSED.
      schema:
        type: string
        format: uuid
    DealStatusFilter:
      name: status
      in: query
      required: false
      description: Optional exact Deal status filter.
      schema:
        $ref: '#/components/schemas/DealStatus'
    Page:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    PageSize:
      name: size
      in: query
      required: false
      description: Requested number of items per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
    DealSort:
      name: sort
      in: query
      required: false
      description: >
        Single allowlisted Deal sort. The public value is not a persistence property
        name. Defaults to newest creation time first.
      schema:
        type: string
        default: createdAt,desc
        enum:
          - createdAt,asc
          - createdAt,desc
          - title,asc
          - title,desc
...
  responses:
    MalformedRequest:
      description: Request JSON cannot be parsed; Problem Details code is MALFORMED_REQUEST.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    ValidationFailed:
      description: Request fields are invalid; Problem Details code is VALIDATION_FAILED.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    SessionRequired:
      description: >
        No active authenticated session exists; Problem Details code is
        AUTH_SESSION_EXPIRED for missing, invalid, or expired sessions.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    LegalEntityAccessDenied:
      description: >
        Required legal entity context is missing, malformed, or does not match the
        requested organization scope; Problem Details code is LEGAL_ENTITY_ACCESS_DENIED.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    LegalEntityNotFoundOrHidden:
      description: >
        Legal entity does not exist or is hidden because the authenticated user is not
        a member; both cases use the same Problem Details code LEGAL_ENTITY_NOT_FOUND.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    DealOrLegalEntityNotFoundOrHidden:
      description: >
        The syntactically valid active legal entity does not exist or is hidden
        because the authenticated user is not a member (LEGAL_ENTITY_NOT_FOUND), or
        the Deal does not exist or is hidden because the authorized active legal
        entity is not a participant (DEAL_NOT_FOUND). Both cases preserve
        non-disclosure at their respective authorization boundary.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    DealScopedMutationForbidden:
      description: >
        CSRF validation failed (CSRF_TOKEN_INVALID), or the required active legal
        entity context header is missing or malformed
        (LEGAL_ENTITY_ACCESS_DENIED), or a visible Deal participant is not the
        Deal's immutable initiator and therefore cannot perform this Deal mutation
        (DEAL_MUTATION_FORBIDDEN). A syntactically valid but nonexistent or hidden
        legal entity returns 404 LEGAL_ENTITY_NOT_FOUND instead.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    DealReviewConflict:
      description: >
        No current REVIEW_REQUIRED analysis is available for review; Problem Details
        code is DEAL_STATE_CONFLICT.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    DealReviewAcceptanceForbidden:
      description: >
        CSRF validation failed (CSRF_TOKEN_INVALID), the required active legal entity
        context is missing or malformed (LEGAL_ENTITY_ACCESS_DENIED), or the visible
        Deal participant is not the immutable Deal initiator
        (DEAL_REVIEW_ACCEPTANCE_FORBIDDEN). Participant visibility never grants
        review acceptance authority.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    DealReviewAcceptanceConflict:
      description: >
        The supplied expectedVersion is stale (DEAL_STALE_VERSION), the target
        analysis is not the current REVIEW_REQUIRED analysis or has been superseded
        (DEAL_STATE_CONFLICT), or Idempotency-Key was reused for a different request
        (IDEMPOTENCY_KEY_REUSED).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    RuleSetVersionNotFoundOrHidden:
      description: >
        The rule-set version does not exist, does not belong to the visible Deal, or
        the Deal is hidden from a non-participant; all cases use
        RULE_SET_VERSION_NOT_FOUND and do not disclose existence.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    CreateDealRequest:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Required title; must be non-blank after trimming and is stored trimmed.
        description:
          type: [string, 'null']
          maxLength: 4000
          description: >-
            Optional Deal description; when supplied it may explicitly be null.
            Blank or whitespace-only values are normalized to null.
...
    DealStatus:
      type: string
      description: Closed authoritative Deal container status set.
      enum: [DRAFT, ACTIVE, CANCELLED, COMPLETED, ARCHIVED]
...
    DealLifecycleProjection:
      type: string
      description: >
        Backend-derived display lifecycle. It is not mutable authoritative state and
        clients must not derive it from status combinations.
      enum:
        - DRAFT
        - CONTRACT_ANALYSIS
        - MANUAL_REVIEW
        - RATIFICATION
        - FUNDING
        - FULFILLMENT
        - SETTLEMENT
        - DISPUTE
        - COMPLETED
        - CANCELLED
        - ARCHIVED
...
    DealExtractionReview:
      type: object
      additionalProperties: false
      description: >
        Participant-readable, immutable review input for the current REVIEW_REQUIRED
        extraction. It does not persist a draft and does not represent ratification.
      required: [analysisId, documentId, rules]
      properties:
        analysisId:
          type: string
          format: uuid
          description: Target analysis identity required by the acceptance action.
        documentId:
          type: string
          format: uuid
        rules:
          type: array
          description: Canonical extracted rules to decide in the acceptance action; never null.
          items:
            $ref: '#/components/schemas/ExtractedRule'
...
    AcceptExtractionReviewRequest:
      type: object
      additionalProperties: false
      required: [analysisId, expectedVersion, decisions]
      properties:
        analysisId:
          type: string
          format: uuid
          description: Current REVIEW_REQUIRED analysis being accepted.
        expectedVersion:
          type: integer
          format: int64
          minimum: 0
          description: Current Deal optimistic-lock version.
        decisions:
          type: array
          description: >
            Complete bulk review decisions. The backend requires exactly one decision
            for every extracted rule and semantic validation allows an empty array
            when the target extraction contains no rules. The backend assigns unique
            manual-N rule references to ADDED decisions within the created
            RuleSetVersion.
          items:
            $ref: '#/components/schemas/ReviewRuleDecision'
...
    RuleSetVersion:
      type: object
      additionalProperties: false
      required: [id, version, sourceAnalysisId, sourceExtractionResultVersionId, createdAt, createdByUserId, previousRuleSetVersionId, ruleCount, rules, excludedRuleReferences]
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
          minimum: 1
        sourceAnalysisId:
          type: string
          format: uuid
          description: Target analysis identity accepted to create this immutable version.
        sourceExtractionResultVersionId:
          type: string
          format: uuid
          description: Immutable canonical ExtractionResultVersion from which this rule set was accepted.
        createdAt:
          $ref: '#/components/schemas/UtcTimestamp'
        createdByUserId:
          type: string
          format: uuid
        previousRuleSetVersionId:
          type: [string, 'null']
          format: uuid
        ruleCount:
          type: integer
          minimum: 0
        rules:
          type: array
          description: Final valid immutable rules only; EXCLUDED rules are absent.
          items:
            $ref: '#/components/schemas/RuleSetRule'
        excludedRuleReferences:
          type: array
          uniqueItems: true
          description: Immutable review-decision history for excluded extracted rules; never null.
          items:
            type: string
            minLength: 1
    RuleSetVersionHistory:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          description: Immutable accepted rule-set versions, including superseded history; never null.
          items:
            $ref: '#/components/schemas/RuleSetVersionSummary'
...
    DealAvailableActions:
      type: object
      additionalProperties: false
      description: >
        Backend-derived action availability for the current user and active legal
        entity. Clients use this projection for presentation but the backend always
        re-authorizes mutations.
      required: [canUpdate, canCancel, canCreateInvitation, canManageParties, canCreateDocumentUploadIntent, canRequestAnalysis]
      properties:
        canUpdate:
          type: boolean
        canCancel:
          type: boolean
        canCreateInvitation:
          type: boolean
          description: >
            True only when the active legal entity is authorized as the DRAFT Deal
            initiator to create invitations.
        canManageParties:
          type: boolean
          description: >
            Backend-derived party-management availability for the current user and
            active legal entity. True only for the authorized DRAFT Deal initiator;
            it does not represent consent authority or Deal activation eligibility.
        canCreateDocumentUploadIntent:
          type: boolean
          description: >
            Backend-derived document upload availability. True only for the Deal
            initiator when its current lifecycle permits a new upload intent;
            participant visibility does not imply this authority.
        canRequestAnalysis:
          type: boolean
          description: >
            Backend-derived analysis-request availability. True only for the Deal
            initiator when the Deal is non-terminal, its current document is AVAILABLE,
            and no analysis job is QUEUED or PROCESSING for that document.
        canReviewExtraction:
          type: boolean
          description: >
            Optional backend-derived review-acceptance availability. When absent or
            unknown, clients must treat it as false and remain read-only. When true,
            the active user is authorized for the Deal initiator and the current
            analysis is REVIEW_REQUIRED.
        canCreateRatificationPackage:
          type: boolean
          description: Optional actor-aware availability; absent or unknown means false/read-only.
        canApproveRatification:
          type: boolean
          description: Optional actor-aware availability; absent or unknown means false/read-only.
        canRejectRatification:
          type: boolean
          description: Optional actor-aware availability; absent or unknown means false/read-only.
        canCreateFundingPlan:
          type: boolean
          description: >
            Optional actor-aware availability. True only for buyer entity ADMIN while
            the Deal is ACTIVE and no FundingPlan exists yet. Absent or unknown means
            false/read-only.
        canInitiateFunding:
          type: boolean
          description: >
            Optional actor-aware availability. True only for buyer entity ADMIN when
            the Deal's single FundingUnit accepts a new payment operation (not FUNDED
            and no CREATED or UNCONFIRMED in-flight operation). Absent or unknown
            means false/read-only.
        canReconcilePaymentOperation:
          type: boolean
          description: >
            Optional actor-aware availability. True only for buyer entity ADMIN when
            the Deal's current PaymentOperation is UNCONFIRMED. Absent or unknown
            means false/read-only.
        canStartFulfillment:
          type: boolean
          description: >
            Optional actor-aware availability. True only for seller entity ADMIN or
            MEMBER when the Deal is ACTIVE and FUNDED and no fulfillment exists yet.
            Absent or unknown means false/read-only.
        canUploadEvidence:
          type: boolean
          description: >
            Optional actor-aware availability. True only for seller entity ADMIN or
            MEMBER when the Deal is ACTIVE and FUNDED and the fulfillment accepts new
            evidence. Absent or unknown means false/read-only.
        canAcceptEvidence:
          type: boolean
          description: >
            Optional actor-aware availability. True only for buyer entity ADMIN when
            the fulfillment has current SUBMITTED evidence awaiting review. Absent or
            unknown means false/read-only.
        canRejectEvidence:
          type: boolean
          description: >
            Optional actor-aware availability. True only for buyer entity ADMIN when
            the fulfillment has current SUBMITTED evidence awaiting review. Absent or
            unknown means false/read-only.
        canOpenDispute:
          type: boolean
          description: >
            Optional actor-aware availability. True only for buyer or seller entity
            ADMIN when the Deal is ACTIVE, fulfillment exists in an eligible opening
            status, and no OPEN or UNDER_REVIEW case exists. Absent or unknown means
            false/read-only.
    UtcTimestamp:
      type: string
      format: date-time
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(?:\.\d+)?Z$'
      description: RFC 3339 timestamp normalized to UTC with a trailing uppercase Z.
...
    DealParticipant:
      type: object
      additionalProperties: false
      description: >
        Deal visibility participant projection. partyRoles describes the current
        buyer/seller assignment only; participation and a party role convey no
        contractual consent or lifecycle authority.
      required: [legalEntityId, legalName, joinedAt, partyRoles]
      properties:
        legalEntityId:
          type: string
          format: uuid
        legalName:
          type: string
          minLength: 1
          maxLength: 200
        joinedAt:
          $ref: '#/components/schemas/UtcTimestamp'
        partyRoles:
          type: array
          minItems: 0
          maxItems: 1
          uniqueItems: true
          description: >
            Current Deal party roles for this participant; [] when unassigned and
            exactly one of BUYER or SELLER when assigned. A participant cannot hold
            both roles.
          items:
            $ref: '#/components/schemas/DealPartyRole'
...
    DealParty:
      type: object
      additionalProperties: false
      description: >
        Current buyer or seller assignment in a Deal detail projection. Its presence
        identifies a role assignment only; it does not indicate ratification,
        contractual consent, or activation.
      required: [legalEntityId, legalName]
      properties:
        legalEntityId:
          type: string
          format: uuid
        legalName:
          type: string
          minLength: 1
          maxLength: 200
    DealDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - reference
        - title
        - description
        - status
        - lifecycle
        - version
        - createdAt
        - updatedAt
        - availableActions
        - buyer
        - seller
        - participants
        - currentDocument
        - analysis
      properties:
        id:
          type: string
          format: uuid
        reference:
          type: string
          minLength: 1
          maxLength: 50
          description: Backend-generated human-readable Deal reference.
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: [string, 'null']
          maxLength: 4000
          description: Always present; null means that no Deal description is currently set.
        status:
          $ref: '#/components/schemas/DealStatus'
        lifecycle:
          $ref: '#/components/schemas/DealLifecycleProjection'
        version:
          type: integer
          format: int64
          minimum: 0
        createdAt:
          $ref: '#/components/schemas/UtcTimestamp'
        updatedAt:
          $ref: '#/components/schemas/UtcTimestamp'
        availableActions:
          $ref: '#/components/schemas/DealAvailableActions'
        buyer:
          description: >
            Current buyer assignment when present; null means no buyer is assigned.
            This field describes assignment only, not contractual consent or Deal
            activation.
          anyOf:
            - $ref: '#/components/schemas/DealParty'
            - type: 'null'
        seller:
          description: >
            Current seller assignment when present; null means no seller is assigned.
            This field describes assignment only, not contractual consent or Deal
            activation.
          anyOf:
            - $ref: '#/components/schemas/DealParty'
            - type: 'null'
        participants:
          type: array
          description: >
            Legal entities with visibility participation in this Deal; never null.
            Pending invitation recipient email is never included.
          items:
            $ref: '#/components/schemas/DealParticipant'
        currentDocument:
          description: >
            Current verified contractual document, or null when no document has been
            finalized. This is a backend-owned pointer; clients must not infer it
            from history or document status.
          anyOf:
            - $ref: '#/components/schemas/AvailableDealDocument'
            - type: 'null'
        analysis:
          $ref: '#/components/schemas/DealDocumentAnalysisSummary'
        currentRuleSet:
          description: >
            Optional backend-owned current accepted rule-set summary. Null means no
            accepted rule-set currently backs this Deal; clients must not infer one
            from history. Its absence is tolerated for additive compatibility.
          anyOf:
            - $ref: '#/components/schemas/RuleSetVersionSummary'
            - type: 'null'
        ratification:
          description: Optional backend-owned ratification readiness and current-package projection.
          anyOf:
            - $ref: '#/components/schemas/RatificationProjection'
            - type: 'null'
        funding:
          description: >
            Optional backend-owned funding summary. Null-tolerant: null means funding
            is not currently projected for this Deal (for example before it is ever
            ACTIVE); absence is tolerated for additive compatibility with clients
            predating Slice 11.
          anyOf:
            - $ref: '#/components/schemas/DealFundingSummary'
            - type: 'null'
        fulfillment:
          description: >
            Optional backend-owned fulfillment summary. Null-tolerant: null means
            fulfillment has not started for this Deal; absence is tolerated for
            additive compatibility with clients predating Slice 12.
          anyOf:
            - $ref: '#/components/schemas/DealFulfillmentSummary'
            - type: 'null'
        casework:
          description: >
            Optional backend-owned active dispute summary for buyer/seller actors only.
            Null-tolerant: null means no active OPEN or UNDER_REVIEW case is currently
            projected for this actor; absence is tolerated for additive compatibility.
            Other participants never receive this field.
          anyOf:
            - $ref: '#/components/schemas/DealCaseworkSummary'
            - type: 'null'
    DealPage:
      type: object
      additionalProperties: false
      required: [items, page, size, totalElements, totalPages]
      properties:
        items:
          type: array
          description: Deal summaries visible to the active legal entity; never null.
          items:
            $ref: '#/components/schemas/DealSummary'
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int32
          minimum: 0
...